Swgoh_TWhelper/requirements.txt
aiohttp==1.0.5
argcomplete==1.8.1
arrow==0.15.5
asgiref==3.2.3
asn1crypto==0.24.0
async-timeout==3.0.0
beautifulsoup4==4.6.0
blinker==1.4
cachetools==2.1.0
certifi==2018.1.18
chardet==3.0.4
checkbox-support==0.22
Click==7.0
cryptography==2.1.4
csvmapper==0.6
cycler==0.10.0
datatables==0.4.9
discord.py==0.16.12
dj-database-url==0.5.0
Django==2.1.15
django-appconf==1.0.3
django-background-tasks==1.2.0
django-compat==1.0.15
django-crispy-forms==1.7.2
django-datatable-view==0.9.0
django-extensions==2.2.1
django-model-utils==4.0.0
django-multiselectfield==0.1.8
django-rest-framework==0.1.0
django-robots==4.0
Django-Select2==5.11.1
django-select2-forms==2.1.0
django-sortedm2m==3.0.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.2
djangorestframework-datatables==0.5.0
djangorestframework-datatables-editor==0.3.2
feedparser==5.2.1
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
guacamole==0.9.2
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.1.1
multidict==4.4.0
netifaces==0.10.4
numexpr==2.6.4
numpy==1.16.2
oauth==1.0.1
oauth2client==4.1.3
oauthlib==2.0.6
padme==1.1.1
pandas==1.0.3
pexpect==4.6.0
Pillow==5.1.0
pipenv==2018.11.26
plainbox==0.25
progressbar==2.5
progressbar2==3.47.0
protobuf==3.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.2
pycairo==1.16.2
pycrypto==2.6.1
pycurl==7.43.0.5
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
python-dateutil==2.6.1
python-debian==0.1.32
python-decouple==3.1
python-gnupg==0.4.1
python-qt-binding==0.3.4
python-utils==2.3.0
pytz==2018.3
pyxdg==0.25
PyYAML==5.3.1
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
rsa==3.4.2
scipy==1.4.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.11.0
sqlparse==0.3.0
ssh-import-id==5.7
systemd-python==234
tables==3.6.1
uritemplate==3.0.0
urllib3==1.22
virtualenv==16.7.2
virtualenv-clone==0.5.3
wadllib==1.3.2
webencodings==0.5
websockets==3.4
xacro==1.13.3
XlsxWriter==0.9.6
xlwt==1.3.0
zope.interface==4.3.2
